requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
idna==3.6
packaging==23.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
djangorestframework-simplejwt
Pillow
